[New] Added test cases for record form form subforms subform traces components trace comparison form
